requirements.txt
pandas
psycopg2
streamlit
SQLAlchemy
pyzmq